split up the patients table view into a feature and into their own components
