+++ client/src/features/patients-table/components/PatientsTableHead.tsx
@@ -1,0 +1,54 @@
+import {
+  Checkbox,
+  TableCell,
+  TableHead,
+  TableRow,
+  TableSortLabel,
+} from "@mui/material";
+import React from "react";
+
+interface EnhancedTableProps {
+  numSelected: number;
+  onSelectAllClick: (event: React.ChangeEvent<HTMLInputElement>) => void;
+  rowCount: number;
+}
+
+type TPatientsTableHead = EnhancedTableProps & {
+  headCells: {
+    id: string;
+    numeric: boolean;
+    disablePadding: boolean;
+    label: string;
+  }[];
+};
+
+export const PatientsTableHead = (props: TPatientsTableHead) => {
+  const { onSelectAllClick, numSelected, rowCount, headCells } = props;
+
+  return (
+    <TableHead>
+      <TableRow>
+        <TableCell padding="checkbox">
+          <Checkbox
+            color="primary"
+            indeterminate={numSelected > 0 && numSelected < rowCount}
+            checked={rowCount > 0 && numSelected === rowCount}
+            onChange={onSelectAllClick}
+            inputProps={{
+              "aria-label": "select all desserts",
+            }}
+          />
+        </TableCell>
+        {headCells.map((headCell) => (
+          <TableCell
+            key={headCell.id}
+            align={headCell.numeric ? "right" : "left"}
+            padding={headCell.disablePadding ? "none" : "normal"}
+          >
+            <TableSortLabel>{headCell.label}</TableSortLabel>
+          </TableCell>
+        ))}
+      </TableRow>
+    </TableHead>
+  );
+};
