+++ client/src/views/ViewPatientsPage.tsx
@@ -1,257 +1,16 @@
 import { useGetAllPatients } from "@client/lib/api/practitioner/query";
 import { useAppStore } from "@client/store";
 import { useSelectPractice } from "@client/store/selectors";
-import { alpha } from "@mui/material/styles";
 import Box from "@mui/material/Box";
-import Table from "@mui/material/Table";
-import TableBody from "@mui/material/TableBody";
-import TableCell from "@mui/material/TableCell";
-import TableContainer from "@mui/material/TableContainer";
-import TableHead from "@mui/material/TableHead";
-import TablePagination from "@mui/material/TablePagination";
-import TableRow from "@mui/material/TableRow";
-import TableSortLabel from "@mui/material/TableSortLabel";
-import Toolbar from "@mui/material/Toolbar";
-import Typography from "@mui/material/Typography";
-import Paper from "@mui/material/Paper";
-import Checkbox from "@mui/material/Checkbox";
-import IconButton from "@mui/material/IconButton";
-import Tooltip from "@mui/material/Tooltip";
-import DeleteIcon from "@mui/icons-material/Delete";
-import FilterListIcon from "@mui/icons-material/FilterList";
-import { usePatientsTable } from "@client/lib/hooks/table/patients-table/usePatientsTable";
-import { usePatientActions } from "@client/lib/hooks/patient/usePatientActions";
-import { DialogConfirmation } from "@client/lib/components/dialog/DialogConfirmation";
+import { PatientsTable } from "@client/features/patients-table/components/PatientsTable";
 
-const headCells = [
-  {
-    id: "select-all",
-    numeric: false,
-    disablePadding: true,
-    label: "Select All",
-  },
-  {
-    id: "name",
-    numeric: true,
-    disablePadding: false,
-    label: "Name",
-  },
-  {
-    id: "email",
-    numeric: true,
-    disablePadding: false,
-    label: "Email",
-  },
-  {
-    id: "address",
-    numeric: true,
-    disablePadding: false,
-    label: "Address",
-  },
-];
-
-interface EnhancedTableProps {
-  numSelected: number;
-
-  onSelectAllClick: (event: React.ChangeEvent<HTMLInputElement>) => void;
-
-  rowCount: number;
-}
-
-function EnhancedTableHead(props: EnhancedTableProps) {
-  const { onSelectAllClick, numSelected, rowCount } = props;
-
-  return (
-    <TableHead>
-      <TableRow>
-        <TableCell padding="checkbox">
-          <Checkbox
-            color="primary"
-            indeterminate={numSelected > 0 && numSelected < rowCount}
-            checked={rowCount > 0 && numSelected === rowCount}
-            onChange={onSelectAllClick}
-            inputProps={{
-              "aria-label": "select all desserts",
-            }}
-          />
-        </TableCell>
-        {headCells.map((headCell) => (
-          <TableCell
-            key={headCell.id}
-            align={headCell.numeric ? "right" : "left"}
-            padding={headCell.disablePadding ? "none" : "normal"}
-          >
-            <TableSortLabel>{headCell.label}</TableSortLabel>
-          </TableCell>
-        ))}
-      </TableRow>
-    </TableHead>
-  );
-}
-interface EnhancedTableToolbarProps {
-  numSelected: number;
-  handleDeletePatientsDialog: (open: boolean) => void;
-}
-
-function EnhancedTableToolbar(props: EnhancedTableToolbarProps) {
-  const { numSelected, handleDeletePatientsDialog } = props;
-  return (
-    <Toolbar
-      sx={[
-        {
-          pl: { sm: 2 },
-          pr: { xs: 1, sm: 1 },
-        },
-        numSelected > 0 && {
-          bgcolor: (theme) =>
-            alpha(
-              theme.palette.primary.main,
-              theme.palette.action.activatedOpacity
-            ),
-        },
-      ]}
-    >
-      {numSelected > 0 ? (
-        <Typography
-          sx={{ flex: "1 1 100%" }}
-          color="inherit"
-          variant="subtitle1"
-        >
-          {numSelected} selected
-        </Typography>
-      ) : (
-        <Typography sx={{ flex: "1 1 100%" }} variant="h6" id="tableTitle">
-          All Patients
-        </Typography>
-      )}
-      {numSelected > 0 ? (
-        <Tooltip title="Delete">
-          <IconButton onClick={() => handleDeletePatientsDialog(true)}>
-            <DeleteIcon />
-          </IconButton>
-        </Tooltip>
-      ) : (
-        <Tooltip title="Filter list">
-          <IconButton>
-            <FilterListIcon />
-          </IconButton>
-        </Tooltip>
-      )}
-    </Toolbar>
-  );
-}
 export const ViewPatientsPage = () => {
   const practice = useAppStore(useSelectPractice);
   const practiceId = String(practice?.id);
 
   const { data: patients } = useGetAllPatients(practiceId);
 
-  const { openDeleteDialog, handleDeletePatientsDialog, handleDeletePatient } =
-    usePatientActions();
+  const renderPatientsTable = <PatientsTable patients={patients} />;
 
-  const {
-    selected,
-    page,
-    dense,
-    rowsPerPage,
-    handleSelectAllClick,
-    handleClick,
-    handleChangePage,
-    handleChangeRowsPerPage,
-    emptyRows,
-    visibleRows,
-  } = usePatientsTable({ patients });
-
-  const renderDeletePatientsDialog = (
-    <DialogConfirmation
-      open={openDeleteDialog}
-      title="Delete Patients"
-      body="Are you sure you want to delete the selected patient(s)? This action cannot be undone."
-      onConfirm={() => handleDeletePatient(selected)}
-      onCancel={() => handleDeletePatientsDialog(false)}
-    />
-  );
-
-  return (
-    <Box sx={{ width: "100%" }}>
-      <Paper sx={{ width: "100%", mb: 2 }}>
-        <EnhancedTableToolbar
-          numSelected={selected.length}
-          handleDeletePatientsDialog={handleDeletePatientsDialog}
-        />
-        <TableContainer>
-          <Table
-            sx={{ minWidth: 750 }}
-            aria-labelledby="tableTitle"
-            size={dense ? "small" : "medium"}
-          >
-            <EnhancedTableHead
-              numSelected={selected.length}
-              onSelectAllClick={handleSelectAllClick}
-              rowCount={patients?.length ?? 0}
-            />
-            <TableBody>
-              {visibleRows.map((row, index) => {
-                if (!row.id) return null;
-                const isItemSelected = selected.includes(row.id);
-                const labelId = `enhanced-table-checkbox-${index}`;
-
-                return (
-                  <TableRow
-                    hover
-                    onClick={(event) => handleClick(event, row.id)}
-                    role="checkbox"
-                    aria-checked={isItemSelected}
-                    tabIndex={-1}
-                    key={row.id}
-                    selected={isItemSelected}
-                    sx={{ cursor: "pointer" }}
-                  >
-                    <TableCell padding="checkbox">
-                      <Checkbox
-                        color="primary"
-                        checked={isItemSelected}
-                        inputProps={{
-                          "aria-labelledby": labelId,
-                        }}
-                      />
-                    </TableCell>
-                    <TableCell
-                      component="th"
-                      id={labelId}
-                      scope="row"
-                      padding="none"
-                    >
-                      {row.name}
-                    </TableCell>
-                    <TableCell align="right">{row.name}</TableCell>
-                    <TableCell align="right">{row.email}</TableCell>
-                    <TableCell align="right">{row.address}</TableCell>
-                  </TableRow>
-                );
-              })}
-              {emptyRows > 0 && (
-                <TableRow
-                  style={{
-                    height: (dense ? 33 : 53) * emptyRows,
-                  }}
-                >
-                  <TableCell colSpan={6} />
-                </TableRow>
-              )}
-            </TableBody>
-          </Table>
-        </TableContainer>
-        <TablePagination
-          rowsPerPageOptions={[5, 10, 25]}
-          count={patients?.length ?? 0}
-          rowsPerPage={rowsPerPage}
-          page={page}
-          onPageChange={handleChangePage}
-          onRowsPerPageChange={handleChangeRowsPerPage}
-        />
-      </Paper>
-      {renderDeletePatientsDialog}
-    </Box>
-  );
+  return <Box sx={{ width: "100%" }}>{renderPatientsTable}</Box>;
 };
